:hammer: refactor: redis_utils 위치 변경 및 services class로 묶음
